# CLI: mask sensitive ENV values