## PR Info

- Closes https://github.com/SeaQL/sea-orm/issues/2656

## `cargo r -- generate entity -h`

- Before:
    ```
    $ cargo r -- generate entity -h
    Compiling sea-orm-cli v1.2.0-rc.1
        Finished `dev` profile [unoptimized + debuginfo] target(s) in 5.38s
        Running `target\debug\sea-orm-cli generate entity -h`
    Generate entity

    Usage: sea-orm-cli generate entity [OPTIONS] --database-url <DATABASE_URL>

    Options:
    ...
    -u, --database-url <DATABASE_URL>
            Database URL [env: DATABASE_URL=mysql://root:root@localhost]
    ...
    ```
- After:
    ```
    $ cargo r -- generate entity -h
    Compiling sea-orm-cli v1.2.0-rc.1
        Finished `dev` profile [unoptimized + debuginfo] target(s) in 5.95s
        Running `target\debug\sea-orm-cli generate entity -h`
    Generate entity

    Usage: sea-orm-cli generate entity [OPTIONS] --database-url <DATABASE_URL>

    Options:
    ...
    -u, --database-url <DATABASE_URL>
            Database URL [env: DATABASE_URL]
    ...
    ```


## `cargo r -- migrate -h`

- Before:
    ```
    $ cargo r -- migrate -h
        Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.27s
        Running `target\debug\sea-orm-cli migrate -h`
    Migration related commands

    Usage: sea-orm-cli migrate [OPTIONS] [COMMAND]

    Options:
    ...
    -u, --database-url <DATABASE_URL>
            Database URL [env: DATABASE_URL=mysql://root:root@localhost]
    ...
    ```
- After:
    ```
    $ cargo r -- migrate -h
    Compiling sea-orm-cli v1.2.0-rc.1
        Finished `dev` profile [unoptimized + debuginfo] target(s) in 22.20s
        Running `target\debug\sea-orm-cli migrate -h`
    Migration related commands

    Usage: sea-orm-cli migrate [OPTIONS] [COMMAND]

    Options:
    ...
    -u, --database-url <DATABASE_URL>
            Database URL [env: DATABASE_URL]
    ...
    ```


## `cargo r -- migrate`

- Before:
    ```
    $ cargo r -- migrate
    Compiling sea-orm-cli v1.2.0-rc.1
        Finished `dev` profile [unoptimized + debuginfo] target(s) in 13.02s
        Running `target\debug\sea-orm-cli migrate`
    Running `cargo run --manifest-path ./migration/Cargo.toml -- up -u mysql://root:root@localhost`
    ```
- After:
    ```
    $ cargo r -- migrate
    Compiling sea-orm-cli v1.2.0-rc.1
        Finished `dev` profile [unoptimized + debuginfo] target(s) in 8.04s
        Running `target\debug\sea-orm-cli migrate`
    Running `cargo run --manifest-path ./migration/Cargo.toml -- up`
    ```